# Speed up parsing of profile files