Read multiple words at once if we know the exact number of words to
read.

Speedup on a 60MiB profile:

PyPy:    1.24s vs 0.84s (1.48x)
CPython: 7.85s vs 3.05s (2.57x)