@jonashaag 2016-05-19
Is there something that I should improve about this PR or is it ready to merge?