@fijal 2016-05-19
I'm +1 on merging it, but I don't like the islice usage - it's unnecessary, unlikely to make things faster and looks quite hackish to me (I have to look up the documentation of islice if I need to read this code). I know people like itertools but personally I think it's not really a solution