@jonashaag 2016-05-19
> I find the loop version much much easier to read for example

I'm getting the impression that you are used to imperative rather than functional thinking... but I don't want to argue coding style here ;) The thing is that we read a large list of `[kind, pc, kind, pc, ...]` and have to extract pairs of two, how would you code that? Like this?

```py
for i in range(len(pcs_and_kinds), 2):
    res.append(wrap_kind(pcs_and_kinds[i], pcs_and_kinds[i+1]))
```