@fijal 2016-05-19
Well, yes, python is a terrible functional programming language :-) I'm generally ok with using such stuff in other languages, but in python without working recursion it's just a PITA. It's also a matter of consistency :-)

or like that: return [wrap_kind(pcs_and_kinds[i], pcs_and_kinds[i + 1]) for i in range(0, len(pcs_and_kinds), 2)]